Support multiple templates folders (#2)
Provide support to append additional templates folders for ionit. With
this change it is not required to run multiple ionit commands per
templates folder, because they could be appended in the command line.

Example:

```
ionit -c tests/config/static \
      -t tests/template/static \
      -t tests/template/static2/
```

This feature is requested to run one ionit service for `/etc` and `/opt`
templates at the same time.